requirements.txt
streamlit
openai
pyswisseph
timezonefinder
pytz
requests
opencv-python
numpy
python-dateutil
